 \docType{data}
    \name{prumetalrates}
    \alias{prumetalrates}
    \title{Rates for prumetal dataset}
    \format{a vector of rates}
    \description{
      This data provides an example vector of rates to the user.
    }
    \author{
      Alex Dornburg
    }
    \usage{prumetalrates}
    \keyword{datasets}