 \docType{data}
    \name{sample.rates}
    \alias{sample.rates}
    \title{Simulated rates}
    \format{a vector of rates}
    \description{
      This data provides an example vector of simulated rates to the user.
    }
    \author{
      Alex Dornburg
    }
    \usage{sample.rates}
    \keyword{datasets}
