 \docType{data}
    \name{sample.tree}
    \alias{sample.tree}
    \title{Sample tree}
    \format{a phy file that can be read in by ape}
    \description{
      This data provides an example tree to the user.
    }
    \author{
      Alex Dornburg
    }
    \usage{sample.tree}
    \keyword{datasets}
