 \docType{data}
    \name{sample.tree2}
    \alias{sample.tree2}
    \title{Another Sample tree}
    \format{a phy file that can be read in by ape}
    \description{
      This data provides a second example tree to the user.
    }
    \author{
      Alex Dornburg
    }
    \usage{sample.tree2}
    \keyword{datasets}
